mac_requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==23.1.0
audio-recorder-streamlit==0.0.8
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.8.0
environs==9.5.0
extra-streamlit-components==0.1.60
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.58.0
gTTS==2.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langchain-experimental==0.0.47
langsmith==0.0.75
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
minio==7.2.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
openai==1.6.1
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
playsound==1.3.0
plotly==5.18.0
protobuf==4.25.1
pyarrow==14.0.2
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.8.0
pymilvus==2.3.4
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
safetensors==0.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
streamlit==1.29.0
streamlit-audiorecorder==0.0.4
streamlit-authenticator==0.2.3
streamlit-pills==0.3.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.5.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.2
tornado==6.4
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
urllib3==2.1.0
validators==0.22.0
yarl==1.9.4
yaspin==3.0.1
zipp==3.17.0
